+++ openstates/data/models/base.py
@@ -1,7 +1,7 @@
 import re
 import uuid
 from django.db import models
-from django.contrib.postgres.fields import ArrayField, JSONField
+from django.contrib.postgres.fields import JSONField
 from django.core.validators import RegexValidator
 from .. import common
 
@@ -52,7 +52,6 @@ class OCDBase(models.Model):
         blank=True,
         help_text="A key-value store for storing arbitrary information not covered elsewhere.",
     )
-    locked_fields = ArrayField(base_field=models.TextField(), blank=True, default=list)
 
     class Meta:
         abstract = True