+++ openstates/scrape/bill.py
@@ -140,27 +140,25 @@ class Bill(SourceMixin, AssociatedLinkMixin, BaseModel):
         self.other_identifiers.append({"identifier": identifier})
 
     def add_document_link(
-        self, note, url, *, date="", media_type="", text="", on_duplicate="warn"
+        self, note, url, *, date="", media_type="", on_duplicate="warn"
     ):
         return self._add_associated_link(
             collection="documents",
             note=note,
             url=url,
             date=date,
-            text=text,
             media_type=media_type,
             on_duplicate=on_duplicate,
         )
 
     def add_version_link(
-        self, note, url, *, date="", media_type="", text="", on_duplicate="warn"
+        self, note, url, *, date="", media_type="", on_duplicate="warn"
     ):
         return self._add_associated_link(
             collection="versions",
             note=note,
             url=url,
             date=date,
-            text=text,
             media_type=media_type,
             on_duplicate=on_duplicate,
         )